+++ tests/test_aact.py
@@ -36,7 +36,7 @@ def _make_conditions() -> pl.DataFrame:
 
 
 def test_source_and_provider():
-    from clinical_mining.data_sources.aact import extract_clinical_report
+    from clinical_mining.provider.aact import extract_clinical_report
 
     result = extract_clinical_report(
         studies=_make_studies(),
@@ -49,7 +49,7 @@ def test_source_and_provider():
 
 
 def test_detailed_descriptions_column_present():
-    from clinical_mining.data_sources.aact import extract_clinical_report
+    from clinical_mining.provider.aact import extract_clinical_report
 
     detailed_descriptions = pl.DataFrame(
         {
@@ -72,7 +72,7 @@ def test_detailed_descriptions_column_present():
 
 
 def test_no_detailed_descriptions_column_absent():
-    from clinical_mining.data_sources.aact import extract_clinical_report
+    from clinical_mining.provider.aact import extract_clinical_report
 
     result = extract_clinical_report(
         studies=_make_studies(),
@@ -84,7 +84,7 @@ def test_no_detailed_descriptions_column_absent():
 
 
 def test_detailed_description_value_preserved():
-    from clinical_mining.data_sources.aact import extract_clinical_report
+    from clinical_mining.provider.aact import extract_clinical_report
 
     detailed_descriptions = pl.DataFrame(
         {
@@ -109,7 +109,7 @@ def test_detailed_description_value_preserved():
 
 def test_replace_with_llm_indications():
     """Test that LLM indications replace source indications for LLM-covered trials."""
-    from clinical_mining.data_sources.aact.clinical_report import (
+    from clinical_mining.provider.aact.clinical_report import (
         replace_with_llm_indications,
     )
 
@@ -228,7 +228,7 @@ def test_replace_with_llm_indications():
 
 
 def test_extract_clinical_report_with_sponsors():
-    from clinical_mining.data_sources.aact import extract_clinical_report
+    from clinical_mining.provider.aact import extract_clinical_report
 
     sponsors = pl.DataFrame(
         {
@@ -269,7 +269,7 @@ def test_extract_clinical_report_with_sponsors():
 
 
 def test_extract_clinical_report_with_study_references():
-    from clinical_mining.data_sources.aact import extract_clinical_report
+    from clinical_mining.provider.aact import extract_clinical_report
 
     study_references = pl.DataFrame(
         {
