+++ tests/test_pubmed.py
@@ -2,7 +2,7 @@
 
 from unittest.mock import MagicMock, patch
 
-from clinical_mining.data_sources.pubmed import (
+from clinical_mining.provider.pubmed import (
     build_publications_map,
     fetch_publications,
 )
@@ -69,10 +69,8 @@ def test_fetch_publications_empty_list():
 def test_fetch_publications_success():
     handle = _mock_entrez_read(TWO_ARTICLES)
     with (
-        patch("clinical_mining.data_sources.pubmed.Entrez.efetch", return_value=handle),
-        patch(
-            "clinical_mining.data_sources.pubmed.Entrez.read", return_value=TWO_ARTICLES
-        ),
+        patch("clinical_mining.provider.pubmed.Entrez.efetch", return_value=handle),
+        patch("clinical_mining.provider.pubmed.Entrez.read", return_value=TWO_ARTICLES),
     ):
         result = fetch_publications([12345678, 99999999])
 
@@ -95,10 +93,8 @@ def test_fetch_publications_missing_abstract_becomes_empty_string():
     }
     handle = _mock_entrez_read(no_abstract)
     with (
-        patch("clinical_mining.data_sources.pubmed.Entrez.efetch", return_value=handle),
-        patch(
-            "clinical_mining.data_sources.pubmed.Entrez.read", return_value=no_abstract
-        ),
+        patch("clinical_mining.provider.pubmed.Entrez.efetch", return_value=handle),
+        patch("clinical_mining.provider.pubmed.Entrez.read", return_value=no_abstract),
     ):
         result = fetch_publications([11111111])
 
@@ -112,9 +108,9 @@ def test_fetch_publications_retries_on_incomplete_read(capsys):
     handle = _mock_entrez_read(SINGLE_ARTICLE)
     efetch = MagicMock(side_effect=[IncompleteRead(b""), handle])
     with (
-        patch("clinical_mining.data_sources.pubmed.Entrez.efetch", efetch),
+        patch("clinical_mining.provider.pubmed.Entrez.efetch", efetch),
         patch(
-            "clinical_mining.data_sources.pubmed.Entrez.read",
+            "clinical_mining.provider.pubmed.Entrez.read",
             return_value=SINGLE_ARTICLE,
         ),
     ):
@@ -149,7 +145,7 @@ def test_build_publications_map_basic():
         "99999999": {"title": "Paper B", "abstractText": "Abstract B"},
     }
     with patch(
-        "clinical_mining.data_sources.pubmed.fetch_publications", return_value=pub_data
+        "clinical_mining.provider.pubmed.fetch_publications", return_value=pub_data
     ):
         result = build_publications_map(records, max_pubs=2)
 
@@ -170,7 +166,7 @@ def test_build_publications_map_respects_max_pubs():
         str(i): {"title": f"Paper {i}", "abstractText": f"Abs {i}"} for i in range(1, 6)
     }
     with patch(
-        "clinical_mining.data_sources.pubmed.fetch_publications", return_value=pub_data
+        "clinical_mining.provider.pubmed.fetch_publications", return_value=pub_data
     ):
         result = build_publications_map(records, max_pubs=3)
 
@@ -195,7 +191,7 @@ def test_build_publications_map_deduplicates_pmids():
         },
     ]
     with patch(
-        "clinical_mining.data_sources.pubmed.fetch_publications", return_value={}
+        "clinical_mining.provider.pubmed.fetch_publications", return_value={}
     ) as mock_fetch:
         build_publications_map(records, max_pubs=3)
